Move Type descriptors to Abstractions assembly
#28 breaking
