+++ src/Violet.WorkItems.Abstractions/Types/ConditionDescriptors.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Violet.WorkItems.Types;
+
+public abstract record ConditionDescriptor(string Type);
+
+public record PropertyValueConditionDescriptor(string PropertyName, string ExpectedValue)
+    : ConditionDescriptor("PropertyValue");
