+++ src/Violet.WorkItems.Abstractions/Types/ValidatorDescriptors.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Violet.WorkItems.Types;
+
+public record ValidatorDescriptor(string Type);
+
+public record StringLengthValidatorDescriptor(int Min, int Max)
+    : ValidatorDescriptor("StringLength");
+public record MandatoryValidatorDescriptor()
+    : ValidatorDescriptor("Mandatory");
