+++ src/Violet.WorkItems.Abstractions/Types/WorkItemDescriptor.cs
