--- src/Violet.WorkItems.Core/Types/Descriptors/ConditionDescriptors.cs
@@ -1,7 +1,0 @@
-using System;
-
-namespace Violet.WorkItems.Types;
-
-public abstract record ConditionDescriptor(string Type);
-
-public record PropertyValueConditionDescriptor(string PropertyName, string ExpectedValue) : ConditionDescriptor("PropertyValue");
