--- src/Violet.WorkItems.Core/Types/Descriptors/ValidatorDescriptors.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace Violet.WorkItems.Types;
-
-public record ValidatorDescriptor(string Type);
-
-public record StringLengthValidatorDescriptor(int Min, int Max) : ValidatorDescriptor("StringLength");
-public record MandatoryValidatorDescriptor() : ValidatorDescriptor("Mandatory");
